Move decode command inside method command
